+++ src/integration-test/java/ca/mcgill/science/tepid/server/auth/AuthIT.kt
@@ -102,7 +102,7 @@ class SessionManagerIT : AuthIT() {
     fun queryUserInDb() {
         val ldapUser = AuthenticationManager.queryUserLdap(PropsLDAPTestUser.TEST_USER)
             ?: fail("Couldn't get test user ${PropsLDAPTestUser.TEST_USER} from LDAP")
-        DB.users.putUser(ldapUser)
+        DB.users.put(ldapUser)
         AuthenticationManager.queryUserDb(PropsLDAPTestUser.TEST_USER)
             ?: fail("User ${PropsLDAPTestUser.TEST_USER} not already in DB")
 
@@ -115,7 +115,7 @@ class SessionManagerIT : AuthIT() {
         val user = AuthenticationManager.queryUser(PropsLDAPTestUser.TEST_USER)
             ?: fail("Couldn't get test user ${PropsLDAPTestUser.TEST_USER} from DB or LDAP")
         user.groups = setOf(AdGroup("DefinitelyFakeGroup"))
-        DB.users.putUser(user)
+        DB.users.put(user)
 
         val refreshedUser = AuthenticationManager.refreshUser(PropsLDAPTestUser.TEST_USER)
         val alteredUser = AuthenticationManager.queryUser(PropsLDAPTestUser.TEST_USER)